Finished cursorary cleanup of unused and duplicate functions.  All code removed has been placed in old/unused.py for reference.
